Latest changes and updates
Changed the dialog with [Y/N] key only easier to use
Re-added Core Isolation and sound tweaks, in case they are needed. This might depend upon the drivers versions used. 
Also, added a warning to rerun the AsMedia Firmware updater, if a change is made to drivers, or if unstability is noticed, especially with USB